AP-14119: Updated SettingsModel to store Strings instead of ints.
This creates more readable configurations and is safer when adding or reordering the underlying enum.
Uses an enum value's name() method which changes only if the enum members are renamed.
In contrast, toString() can be used to display a readable representation for the enum in a UI.

AP-14119 (Create a SettingsModel for Enum Types)